Code revamp
The code on the gui.py program is partially really awful  and everything is in a single file which makes modifying the program a pain in the ass, so I'm modifying the program to use modules instead of having everything in a single file. The changes on this branch will be integrated to the master and release branches when this stuff is tested.